requirements.txt
boto3>=1.35.0
PyYAML>=6.0.1
requests>=2.32.0
beautifulsoup4>=4.12.3
markdownify>=0.13.1
